requirements.txt
beautifulsoup4==4.11.1
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
greenlet==2.0.1
idna==3.4
lxml==4.9.2
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
